moved fake_trades, into trade_classes
Could still use some adjusting by moving the fake_trade modules into Trade
and then using self, code is a looking a bit spaggetti as is

+++ utility.py
@@ -303,7 +303,7 @@ def binance_timestamp_local(timestamp):
     utctime = datetime.utcfromtimestamp(dt)
     timedelta = 60 * 60 * 8  # + 8 Hours from UTC
     timestamp = datetime.timestamp(utctime) + timedelta
-    return timestamp * 1000
+    return int(timestamp * 1000)
 
 
 class Sleeper:
